requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
groq==0.4.1
pydantic==2.5.0
supabase==2.9.1
httpx==0.27.0
python-dotenv==1.0.0
requests==2.32.3